# typescript/packages/apps/dataAsset/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DF: Asset Management",
    "description": "\nHas accountability for:\n- managing Assets in data zone\n\t- creating data brew jobs for profiling and data conversion \n\t- running data bre job for profiling and data quality\n",
    "version": "0.0.1"
  },
  "components": {
    "securitySchemes": {
      "platformUserPool": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "properties": {
          "id": {
            "description": "the identifier of the Amazon DataZone inventory asset..",
            "type": "string"
          },
          "domainId": {
            "description": "The identifier of the Amazon DataZone domain in which the inventory asset exists.",
            "type": "string"
          },
          "name": {
            "description": "The identifier of the Amazon DataZone domain in which the inventory asset.",
            "type": "string"
          },
          "description": {
            "description": "The description of an Amazon DataZone inventory asset.",
            "type": "string"
          },
          "owningProjectId": {
            "description": "The identifier of the Amazon DataZone project that owns the inventory asset.",
            "type": "string"
          },
          "typeIdentifier": {
            "description": "The identifier of the asset type of the specified Amazon DataZone inventory asset.",
            "type": "string"
          },
          "typeRevision": {
            "description": "The revision of the inventory asset type.",
            "type": "string"
          },
          "createdAt": {
            "description": "The timestamp of when the Amazon DataZone inventory asset was created.",
            "type": "string"
          },
          "createdBy": {
            "description": "The Amazon DataZone user who created the inventory asset.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "owningProjectId",
          "typeIdentifier",
          "typeRevision"
        ],
        "title": "dataAssetResource"
      },
...
    }
  },
  "paths": {
...
    "/domains/{domainId}/projects/{projectId}/dataAssets": {
      "get": {
        "tags": [
          "Data Assets"
        ],
        "description": "Lists of data assets.",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "count",
            "required": false,
            "description": "Count of results to return."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "fromToken",
            "required": false,
            "description": "Token used to paginate from (exclusive)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "domainId",
            "required": true,
            "description": "Unique id."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true,
            "description": "Unique id."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "accept-version",
            "required": true,
            "description": "API version"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "accept",
            "required": true,
            "description": "Accepted Content Type"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Success.",
                  "type": "object",
                  "properties": {
                    "dataAssets": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "count": {
                          "description": "No. of results returned when pagination requested.",
                          "type": "integer"
                        },
                        "lastEvaluatedToken": {
                          "description": "Token used to paginate to the next page of search result.",
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "dataAssets"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "security": [],
...
}